Trekkjer ut ansvaret for distribuering til si eiga klasse
