+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/BrevService.kt
@@ -7,9 +7,6 @@ import no.nav.etterlatte.brev.brevbaker.BrevbakerService
 import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode
 import no.nav.etterlatte.brev.brevbaker.LanguageCode
 import no.nav.etterlatte.brev.db.BrevRepository
-import no.nav.etterlatte.brev.distribusjon.DistribusjonService
-import no.nav.etterlatte.brev.distribusjon.DistribusjonsTidspunktType
-import no.nav.etterlatte.brev.distribusjon.DistribusjonsType
 import no.nav.etterlatte.brev.dokarkiv.DokarkivService
 import no.nav.etterlatte.brev.hentinformasjon.BrevdataFacade
 import no.nav.etterlatte.brev.hentinformasjon.SakService
@@ -38,7 +35,6 @@ class BrevService(
     private val soekerService: SoekerService,
     private val adresseService: AdresseService,
     private val dokarkivService: DokarkivService,
-    private val distribusjonService: DistribusjonService,
     private val brevbakerService: BrevbakerService,
     private val brevDataFacade: BrevdataFacade,
 ) {
@@ -191,34 +187,6 @@ class BrevService(
         logger.info("Brev med id=${brev.id} markert som journalført")
 
         return response.journalpostId
-    }
-
-    fun distribuer(id: BrevID): String {
-        val brev = hentBrev(id)
-
-        if (brev.status != Status.JOURNALFOERT) {
-            throw IllegalStateException(
-                "Forventet status ${Status.JOURNALFOERT} på brev (id=${brev.id}), men fikk ${brev.status}",
-            )
-        }
-
-        val journalpostId =
-            requireNotNull(db.hentJournalpostId(id)) {
-                "JournalpostID mangler på brev (id=${brev.id}, status=${brev.status})"
-            }
-
-        val mottaker =
-            requireNotNull(brev.mottaker) {
-                "Mottaker må være satt for å kunne distribuere brevet (id: $id)"
-            }
-
-        return distribusjonService.distribuerJournalpost(
-            brevId = brev.id,
-            journalpostId = journalpostId,
-            type = DistribusjonsType.ANNET,
-            tidspunkt = DistribusjonsTidspunktType.KJERNETID,
-            adresse = mottaker.adresse,
-        )
     }
 
     private fun sjekkOmBrevKanEndres(brevID: BrevID) {
