+++ apps/etterlatte-brev-api/src/test/kotlin/no/nav/etterlatte/brev/BrevRouteTest.kt
@@ -53,6 +53,7 @@ internal class BrevRouteTest {
     private val mockOAuth2Server = MockOAuth2Server()
     private lateinit var hoconApplicationConfig: HoconApplicationConfig
     private val brevService = mockk<BrevService>()
+    private val brevdistribuerer = mockk<Brevdistribuerer>()
     private val tilgangssjekker = mockk<Tilgangssjekker>()
 
     @BeforeAll
@@ -253,6 +254,7 @@ internal class BrevRouteTest {
             restModule(this.log, routePrefix = "api") {
                 brevRoute(
                     brevService,
+                    brevdistribuerer,
                     tilgangssjekker,
                 )
             }
